tests/wpt/web-platform-tests/fullscreen/model/move-to-iframe-manual.html

<!DOCTYPE html>
<title>Move the fullscreen element to another document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div></div>
<iframe></iframe>
<script>
async_test(t => {
  t.add_cleanup(() => {
    Promise.resolve(document.exitFullscreen()).catch(() => {});
  });
  const div = document.querySelector("div");
  const iframe = document.querySelector("iframe");
  document.onfullscreenchange = t.step_func(event => {
    assert_equals(document.fullscreenElement, div);
    assert_equals(event.target, div);

    assert_equals(div.ownerDocument, document);
    iframe.contentDocument.body.appendChild(div);
    assert_not_equals(div.ownerDocument, document);
    // Moving /div/ removes it from the top layer and thus the fullscreen
    // element synchronously becomes null.
    assert_equals(document.fullscreenElement, null);

    div.onfullscreenchange = t.unreached_func("fullscreenchange fired on element");
    iframe.contentDocument.onfullscreenchange = t.unreached_func("fullscreenchange fired on other document");
    document.onfullscreenchange = t.step_func_done(event => {
      assert_equals(document.fullscreenElement, null);
      assert_equals(event.target, document);
    });
  });
  trusted_request(t, div);
});
</script>
