third_party/WebKit/LayoutTests/external/wpt/fullscreen/model/move-to-iframe-manual-expected.txt

This is a testharness.js-based test.
FAIL Move the fullscreen element to another document assert_equals: expected Element node <div></div> but got Document node with 2 children
Harness: the test ran to completion.
